requirements.txt
langchain
langchain_experimental
langchain_core
google-generativeai
google-ai-generativelanguage
langchain-google-genai
docarray
PyPDF2
python-dotenv
streamlit